% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanline.R
\name{scanline}
\alias{scanline}
\title{Convert an image to a retro-futuristic scanline image}
\usage{
scanline(
  image,
  n_scanlines = 60,
  scanline_col = c("black", "darkslategrey", "darkseagreen2", "paleturquoise"),
  opacities = c(1, 0.6, 0, 0, 0.6, 1),
  normalise = FALSE,
  border_size = 1,
  border_intensity = 255,
  frame_size = 10,
  vertical_filter = "Point",
  horizontal_filter = "Triangle",
  composite_operator = "HardLight",
  debug = FALSE,
  print_details = FALSE,
  add_noise = FALSE,
  noise_type = "gaussian"
)
}
\arguments{
\item{image}{a magick image object or image file path/URL}

\item{n_scanlines}{the number of scanlines in the image (default = 60)}

\item{scanline_col}{character vector of colours from dark to light (reverse order for negative image)}

\item{opacities}{vector of opacities. The image will be expanded \code{length(opacities)} times in x and y (default = c(1, 0.6, 0, 0, 0.6, 1))}

\item{normalise}{normalise the image with \code{magick::image_normalise()} before applying scanlines (default = FALSE)}

\item{border_size}{size of border around image in units of scanlines (default = 1)}

\item{border_intensity}{intensity of border (from 0 (dark) to 255 (light)) (default = 255)}

\item{frame_size}{size of frame around image as a percentage of the number of \code{n_scanlines} (default = 10)}

\item{vertical_filter}{resize filter for vertical expansion from \code{magick::filter_types()} (default = "Point")}

\item{horizontal_filter}{resize filter for horizontal expansion from \code{magick::filter_types()} (default = "Triangle")}

\item{composite_operator}{blending operator for scanlines from \code{magick::compose_types()} (default = "HardLight")}

\item{debug}{return an output that contains debugging details (best run with small values for \code{n_scanlines}) (default = FALSE)}

\item{print_details}{print information during processing (default = FALSE)}

\item{add_noise}{add noise to image (default = FALSE)}

\item{noise_type}{\code{magick::noise_types()} value to add to image if \code{add_noise = TRUE} (default = "gaussian")}
}
\description{
Inspired by the aesthetics of the Alien films.
Give an image that old school computer terminal scanline feel.
This is just for fun \strong{and is experimental}. Images may not be rendered correctly and detail will certainly be missing!
}
\details{
Image processing is handled by the {magick} package
}
