% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanline_gif.R
\name{scanline_gif}
\alias{scanline_gif}
\title{Create scanline animation from a GIF}
\usage{
scanline_gif(
  gif,
  width = 600,
  height = 600,
  scale = 1,
  fps = 10,
  delay = NULL,
  background_col = "black",
  ...
)
}
\arguments{
\item{gif}{path to GIF file}

\item{width}{output width (default = 600)}

\item{height}{output height (default = 600)}

\item{scale}{multiplicative scale factor applied to width and height (default = 1)}

\item{fps}{frames per second (passed to \code{magick::image_animate()}) (default = 10)}

\item{delay}{delay after each frame, in 1/100 seconds. Must be length 1, or number of frames.
If specified, then fps is ignored.passed to \code{magick::image_animate()}.}

\item{background_col}{background colour for graphics device (default = "black")}

\item{...}{named arguments passed on to \code{scanline::scanline()}}
}
\description{
Inspired by the aesthetics of the Alien films.
Give an image that old school computer terminal scanline feel.
This is just for fun \strong{and is experimental}. Images may not be rendered correctly and detail will certainly be missing!
}
